backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Authentication & Security
PyJWT==2.8.0
argon2-cffi==23.1.0
python-dotenv==1.0.1
cryptography==42.0.2
pyotp==2.9.0

# AWS & Storage
boto3==1.34.34
django-storages==1.14.2

# Celery & Redis
celery==5.3.6
redis==5.0.1
django-redis==5.4.0

# Email
django-anymail==10.2

# API & Serialization
drf-spectacular==0.27.1
djangorestframework-simplejwt==5.3.1

# Utils
python-dateutil==2.8.2
Pillow==10.2.0
requests==2.31.0

# Development
django-debug-toolbar==4.3.0
ipython==8.21.0

# Testing
pytest==8.0.0
pytest-django==4.7.0
factory-boy==3.3.0

# Monitoring & Logging
sentry-sdk==1.40.3

# CORS & Security
django-csp==3.8
